// src/app/core/api/model/postMetrics.ts
/**
 * Trendors
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CampaignInfluencerPost } from './campaignInfluencerPost';


export interface PostMetrics { 
    /**
     * Unique identifier for the entity
     */
    id?: number;
    /**
     * Measured influencer post
     */
    campaignInfluencerPost?: CampaignInfluencerPost;
    views?: number;
    likes?: number;
    comments?: number;
    shares?: number;
    /**
     * When the numbers were pulled
     */
    recorded_at?: string;
}